requirements.txt
astroid==2.3.2
Click==7.0
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.9.0
pylint==2.4.3
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
